Push recent changes
Change screen manager to use a cache instead of an array. Now, all screens are cached, but up to a limit of 20 total, and for a maximum of ten minutes. This should make it easy to go back and forth between screens quickly (such as the town), and then alleviate pressure when not needed (An extended battle for example, that is contained entirely in the battle screen.)
Change character images on profile screen to scale properly
Use a StackLayout (First time I've used!) to implement the skill tree screen
Implement the PerkInfo popup
Fix Scaling issue for filled character previews
Change Sortable and Filterable from PPCHoverPathButtons to PPCHoverPathToggleButtons, adjust sortable callbacks to correctly update ascending/descending
Change filterable to use layered images instead of tons of images. Adjust filterable to have correct alignment for buttons
Fix square character preview to use get_by_index and proper icons names
Worked on boss encounter animation